Fixed modulus bug

+++ Expr.hs
@@ -2,6 +2,7 @@ module Expr where
 
 import Parsing
 import Data.Maybe
+import Data.Fixed (mod')
 
 
 type Name = String
@@ -18,7 +19,7 @@ data Expr = Add Expr Expr
           | Abs Expr
           | Mod Expr Expr
           | Power Expr Expr
-        
+
   deriving Show
 
 -- These are the REPL commands - set a variable name to a value, and evaluate
@@ -31,12 +32,12 @@ data Command = Set Name Expr
   deriving Show
 
 
-retrieveVar :: [(Name, Int)] -> Name -> Maybe Int 
+retrieveVar :: [(Name, Float)] -> Name -> Maybe Float
 retrieveVar [] x = Nothing
 retrieveVar ((a, b) : vs) x =  if a == x
                                then Just b
                                else retrieveVar vs x
- 
+
 eval :: [(Name, Float)] -> -- Variable name to value mapping
         Expr -> -- Expression to evaluate
         Maybe Float -- Result (if no errors such as missing variables)
@@ -47,7 +48,7 @@ eval vars (Add x y) = do q <-  (eval vars x)
                          p <-  (eval vars y)
                          Just (q + p) -- Just ((eval vars x) + (eval vars y)) --fmap sum $ sequence [eval vars x, eval vars y]  -- return an error (because it's not implemented yet!)
 
-eval vars (Minus x y) =  do q <-  (eval vars x) 
+eval vars (Minus x y) =  do q <-  (eval vars x)
                             p <-  (eval vars y)
                             Just (q - p)
 
@@ -64,11 +65,11 @@ eval vars (Abs x) = do q <-  (eval vars x)
 
 eval vars (Mod x y) = do q <-  (eval vars x)
                          p <-  (eval vars y)
-                         Just (mod' q p ) 
+                         Just (mod' q p)
 
 eval vars (Power x y) = do q <-  (eval vars x)
                            p <-  (eval vars y)
-                           Just (q ** p)                        
+                           Just (q ** p)
 
 digitToInt :: Char -> Int
 digitToInt x = fromEnum x - fromEnum '0'
@@ -84,14 +85,14 @@ pCommand = do string ":q"
               return (Quit)
             ||| do string ":r"
                    f <- filename
-                   return (Read f)      
+                   return (Read f)
                 ||| do t <- identifier
                        char '='
                        e <- pExpr
                        return (Set t e)
                     ||| do e <- pExpr
                            return (Eval e)
-                    
+
 
 pExpr :: Parser Expr
 pExpr = do t <- pTerm
@@ -114,7 +115,7 @@ pFactor = do d <- float
                             e <- pExpr
                             char ')'
                             return e
-                          ||| do char '|' 
+                          ||| do char '|'
                                  e <- pExpr
                                  char '|'
                                  return (Abs e)
@@ -127,12 +128,12 @@ pTerm = do f <- pFactor
             ||| do char '/'
                    t <- pTerm
                    return (Division f t)
-                 ||| do char '^' 
+                 ||| do char '^'
                         t <- pTerm
                         return (Power f t)
-                      ||| do char '%' 
+                      ||| do char '%'
                              t <- pTerm
                              return (Mod f t)
                             ||| return f
 
-                    
+
